# Convert transform to run in browser