--- a/.prettierignore
+++ b/.prettierignore
@@ -5,3 +5,5 @@
 dist/*
 docs/*
 test-coverage/*
+**/*.xml
+**/*.xsl