--- a/config/build.shared.d.ts
+++ b/config/build.shared.d.ts
@@ -1,7 +1,9 @@
-export {
-    config,
-    external,
-    resolvePath,
-    readFile,
-    rootDir,
-} from './build.shared';
+import type { UserConfig } from 'vitest/config';
+
+export const baseConfig: UserConfig;
+export const config: typeof import('./config.json');
+export const define: Record<string, string>;
+export const external: string[];
+export const resolvePath: (path: string) => string;
+export const readFile: (path: string) => string;
+export const rootDir: string;