--- a/src/api.ts
+++ b/src/api.ts
@@ -11,7 +11,7 @@
 
 import express from 'express';
 import { request } from 'undici';
-import { transform } from './transformer';
+import { transform } from './node';
 
 const router = express.Router();