--- a/src/app.ts
+++ b/src/app.ts
@@ -1,7 +1,7 @@
 import bodyParser from 'body-parser';
 import express from 'express';
-import { config } from './config/build.shared';
-import { api } from './src/api';
+import { config } from '../config/build.shared';
+import { api } from './api';
 
 const app = express();
 
@@ -17,5 +17,11 @@ app.use(
 );
 
 api(app);
+
+if (ENV === 'production') {
+    app.listen(app.get('port'), () => {
+        console.warn(`enketo-transformer running on port ${app.get('port')}!`);
+    });
+}
 
 export { app };