--- a/tools/shared.d.ts
+++ b/tools/shared.d.ts
@@ -1,0 +1,3 @@
+export const rootDir: string;
+
+export const resolvePath: (path: string) => string;