… .eslintignore
… .eslintrc.json
… .github/workflows/npmjs.yml
… .vscode/launch.json
… .vscode/settings.json
… README.md
… app.js
… demo/index.html
… demo/package.json
… demo/public/icon.png
… demo/tsconfig.json
… index.html
… package.json
… public/icon.png
… src/node.ts
… src/transformer.ts
… src/url.ts
… src/web.ts
… src/xsl/openrosa2html5form.xsl
… src/xsl/openrosa2xmlmodel.xsl
… test/bad-external.spec.ts
… test/benchmarks.ts
… test/external-fixtures/enketo-core/SOAR_facilitysurvey_full_2017May11_Kenya.xml
… test/external-fixtures/enketo-core/autocomplete-cascade.xml
… test/external-fixtures/enketo-core/calc-control.xml
… test/external-fixtures/enketo-core/calc-in-group-with-relevant.xml
… test/external-fixtures/enketo-core/calcs-cascade.xml
… test/external-fixtures/enketo-core/calcs.xml
… test/external-fixtures/enketo-core/calcs_in_repeats.xml
… test/external-fixtures/enketo-core/calcs_in_repeats_2.xml