## Why?

This is a **big** change. Before I go into detail about what's inside, I want to discuss the reasons I think it's worth doing.

### Performance

Transformer has been a known performance-sensitive part of the Enketo stack for some time, with mitigations like server-side caching in Enketo Express and dev-side caching in Enketo Core to reduce the impact of slow transforms. While this change itself does not address all known performance issues, it does significantly improve the speed of transforms. This improvement is especially noticeable on forms which I frequently use to evaluate performance issues.

Starting with a spike investigating whether the change was even feasible, and having noticed the improvement, I went back and added benchmarks (with a handy summary reporter in GitHub Actions which will hopefully show up in a comment below when I open this PR) so I could show the before/after.

[Before](https://github.com/eyelidlessness/enketo-transformer/actions/runs/3970238972#summary-10785535023) any browser-related changes:

<table role="table"><tbody><tr><th></th><th>Name</th><th>Average</th><th>Worst</th><th>Best</th><th>Samples</th></tr><tr><td>Slowest</td><td>va_who_v1_5_3.xml</td><td>6529.23ms</td><td>6654.23ms</td><td>6393.52ms</td><td>10</td></tr><tr><td>Fastest</td><td>missing-external-instance.xml</td><td>8.81ms</td><td>9.06ms</td><td>8.68ms</td><td>10</td></tr><tr><td>Average overall</td><td colspan="5">69.75ms
</td></tr><tr><td>Total runtime</td><td colspan="5">104.38s
</td></tr></tbody></table>

[After](https://github.com/eyelidlessness/enketo-transformer/actions/runs/4019981081#summary-10916665598):

<table role="table"><tbody><tr><th></th><th>Name</th><th>Average</th><th>Worst</th><th>Best</th><th>Samples</th></tr><tr><td>Slowest</td><td>va_who_v1_5_3.xml</td><td>2936.86ms</td><td>3207.54ms</td><td>2738.79ms</td><td>10</td></tr><tr><td>Fastest</td><td>repeat-calc.xml</td><td>9.13ms</td><td>9.48ms</td><td>8.80ms</td><td>10</td></tr><tr><td>Average overall</td><td colspan="5">42.26ms
</td></tr><tr><td>Total runtime</td><td colspan="5">63.76s
</td></tr></tbody></table>

For XForms which already transform quickly, this difference will probably be unnoticeable. But for XForms which transform slowly, there's potential for up to a 50% reduction in time to transform.

There's obviously still more room for performance improvement! To that point...

### Development and maintenance improvements

The performance delta shown above is from GitHub Actions, on their shared hardware in a virtualized environment. Surprisingly, the performance improvement is even greater in my local dev environment[^1]. These improvements will also allow for quicker iterations both in Transformer and other parts of the Enketo stack.

Combined with the prior change converting Transformer to TypeScript, this change will allow both debugging and visualizing transforms in browser, and adds significant coverage to catch potential regressions (more on that below). This will help development move not just quicker but more safely.

This will also allow us to remove the `libxslt` dependency and its own dependency on `node1-libxmljs-myh` (a fork of `libxmljs`). Besides being a large reason for the performance improvements, this will also improve build times significantly, allowing faster and more flexible iteration. And their removal also eliminates some maintenance burdens/risks—for instance, we'll finally be able to upgrade NPM, and more freely upgrade Node as well.

### Flexibility for downstream improvements

When this change is merged, projects using Enketo Transformer will immediately have the _opportunity_ to run transforms in a browser. Further performance improvements may allow us to eliminate the need for server-side caching in projects like Enketo Express.

## What's in this change

1. The parts of the transform which previously targeted `libxslt` and `libxmljs` have been converted to standard DOM and XSLT APIs, so transforms can be performed in a browser. Part of this change moves a few specific portions of the transform previously implemented in `openrosa2html5form.xsl`—those which depended on XSLT extensions not available in browsers—to DOM code.

2. To support backwards compatibility:

   - Transforms in Node utilize a headless instance of Firefox, controlled by the [`playwright`](https://playwright.dev/) library.

   - Support for the `preprocess` option, which depends on `libxmljs`, is preserved in Node by preprocessing before calling into the headless browser.

3. A simple demo app is added which will show the transform results for all included test fixtures.

## How this change has been tested and validated

My goal from the beginning of this effort was to make every effort to ensure that behavior is functionally identical for users (performance improvements notwithstanding), i.e. both feature- and bug-compatible.

Supporting this effort, test coverage has been expanded to include snapshot tests of the transform results for existing test fixtures under `/test/forms`, as well as the corresponding fixtures from Enketo Core (under `/test/external-fixtures/enketo-core`). These snapshot tests help ensure with very high confidence that nothing has meaningfully changed[^2] in transform behavior.

For the portions of code converted from XSLT to DOM, the implementations are written to mirror the original XSLT as closesly as possible, from naming parity to logical flow.

The project has been installed locally in enketo-core and enketo-express, to verify that all tests pass (see known issues), and that all flows known to depend on Transformer work as expected.

The browser demo also serves as a validation that the new build setup works in a web project.

## Known issues

### Timer mocks (in test environments)

With all of the validation described above, there is currently one issue I identified running tests in Enketo Express: using timer mocks in a test environment can cause failures or an infinite loop if you attempt to run all outstanding timers (e.g. Sinon's `timers.runAll`). This is the case in Express, and its failure causes a cascade of others.

We decided to open this PR now knowing the issue exists, and investigate whether it can be addressed in a later update. We also believe that the use of timers is an implementation detail, not an explicit feature or API contract. But if this is going to be a problem for users outside of Enketo Express, it would be great to get feedback about that!

[^1]: My strong hunch is this is a difference between platforms (Ubuntu and macOS, respectively).

[^2]: By "meaningful", I mean that either the transform results have not changed in any way beyond changes to attribute order or insignificant whitespace.