@lognaturel 2023-01-27
For fun:

| platform                                                           | WHO VA   | SOAR    |
|--------------------------------------------------------------------|----------|---------|
| Firefox, MacBookPro18,3                                            | 1,100ms  | 173ms   |
| Safari, MacBookPro18,3                                             | 4,834ms  | 1,125ms |
| [Moto G5](https://www.gsmarena.com/motorola_moto_g5_plus-8453.php) | 43,358ms | 8,857ms |
| iPhone 13                                                          | 4,760ms  | 1,090ms |