@MartijnR 2023-01-27
Wow, impressive results and a very exciting step forward!

Are you considering switching the headless browser (for PDF generation) in Enketo Express to Firefox as well, or have both FF and Chrome running headlessly?