@eyelidlessness 2023-01-27
> [Moto G5](https://www.gsmarena.com/motorola_moto_g5_plus-8453.php)	43,358ms	8,857ms

Whoa! Now I know which device I need for performance testing.