@eyelidlessness 2023-01-27
> Wow, impressive results and a very exciting step forward!

I’m excited that you’re excited!

> Are you considering switching the headless browser (for PDF generation) in Enketo Express to Firefox as well, or have both FF and Chrome running headlessly?

I hadn’t thought that far along, but this reminds me that I forgot to clarify why headless Firefox was chosen for this!

Before I even started working on a headless browser solution for Node compatibility, I started with Firefox because MDN has extensive documentation on its XSLT support. By the time I got it mostly working, I noticed it performed much better in Firefox than the others. But that’s almost entirely time spent in the XSLT transform. Other than that perf is basically equivalent across the major engines.

To the extent we’re using a headless browser for anything, it would definitely be good to unify on one set of tooling! I noticed Puppeteer does have partial Firefox support, and a mostly compatible API, but chose Playwright mostly because support was more mature for this use case. But I bet it’d be a trivial change in Express, and worth it for minimizing differences and dependency overlap.