@MartijnR 2023-01-27
> To the extent we’re using a headless browser for anything, it would definitely be good to unify on one set of tooling!

Great. FYI, OpenClinica has been running into performance issue with headless browser usage in their Express fork (PDF generation of very large sets of records, and also validating records headlessly which is a custom feature they have). I started playing around with keeping a single puppeteer instance alive (and automatically restarting it if it shuts down for some reason) and headless browser caching here: https://github.com/OpenClinica/enketo-express-oc/pull/587/files (but this is largely untested so far).