@eyelidlessness 2023-01-27
> I started playing around with keeping a single puppeteer instance alive (and automatically restarting it if it shuts down for some reason)

This is a really good point that I should have considered with the persistent browser instance. I'm going to add some inline comments in a bit, I'll make sure to capture that too.